requirements.txt
bcrypt==3.1.6
blinker==1.4
cffi==1.12.2
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==6.0.0
pycparser==2.19
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
WTForms==2.2.1
